refactor(croquis): split script_parser/process into cohesive modules
